# pygmt.grdtrack: Add 'output_type' parameter for output in pandas/numpy/file formats